Refactor objective-c and javascript code to separate generic test code - also setting up refresh token in test case super class (not in AppDelegate)
