+++ .gitignore
@@ -20,3 +20,4 @@ node_modules/
 .github/
 .idea/
 package-lock.json
+npm-debug.log
