+++ libs/SalesforceReact/SalesforceReactTests/ReactTestCase.h
@@ -1,0 +1,60 @@
+/*
+ Copyright (c) 2018-present, salesforce.com, inc. All rights reserved.
+ 
+ Redistribution and use of this software in source and binary forms, with or without modification,
+ are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright notice, this list of conditions
+ and the following disclaimer.
+ * Redistributions in binary form must reproduce the above copyright notice, this list of
+ conditions and the following disclaimer in the documentation and/or other materials provided
+ with the distribution.
+ * Neither the name of salesforce.com, inc. nor the names of its contributors may be used to
+ endorse or promote products derived from this software without specific prior written
+ permission of salesforce.com, inc.
+ 
+ THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR
+ IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND
+ FITNESS FOR A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+ CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL
+ DAMAGES (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
+ DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY,
+ WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY
+ WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+
+// NB: we shouldn't have to define RCT_EXTERN here
+// It is already defined in RCTDefines.h which is imported in RCTBridge.h which is imported in RCTTestRunner.h
+// But somehow the compiler gets confused and won't build
+#if defined(__cplusplus)
+#define RCT_EXTERN extern "C" __attribute__((visibility("default")))
+#define RCT_EXTERN_C_BEGIN extern "C" {
+#define RCT_EXTERN_C_END }
+#else
+#define RCT_EXTERN extern __attribute__((visibility("default")))
+#define RCT_EXTERN_C_BEGIN
+#define RCT_EXTERN_C_END
+#endif
+
+#import <XCTest/XCTest.h>
+#import <RCTTest/RCTTestRunner.h>
+
+#define RCT_TEST(name)                        \
+- (void)test##name                            \
+{                                             \
+    [self.runner runTest:_cmd module:@#name]; \
+}
+
+#define RCT_TEST_FAKE_FAILURE(name)                        \
+- (void)test##name                            \
+{                                             \
+[self.runner runTest:_cmd module:@#name initialProps:nil configurationBlock:nil expectErrorRegex:@"FAKE_FAILURE"]; \
+}
+
+
+@interface ReactTestCase : XCTestCase
+
+@property (nonatomic, strong) RCTTestRunner* runner;
+@property (nonatomic, strong) NSString *jsSuitePath;
+
+@end
+
